Refactor: Replace FinalPopupScanner with UltimateModernXSSScanner
This commit replaces the old FinalPopupScanner with the new UltimateModernXSSScanner. The new scanner offers enhanced features for modern web applications and improved screenshot capabilities.
